rework routing; delete a question

+++ app/routes/index.js
@@ -12,19 +12,26 @@ export default Ember.Route.extend({
       this.transitionTo('index');
     },
 
-    update(question, params) {
-      Object.keys(params).forEach(function(key) {
-        if(params[key]!==undefined) {
-          question.set(key,params[key]);
-        }
-      });
-      question.save();
-      this.transitionTo('index');
-    },
-
-    destroyQuestion(question) {
-      question.destroyRecord();
-      this.transitionTo('index');
+    delete(question) {
+      if (confirm('Are you sure you want to delete this question?')) {
+        question.destroyRecord();
+        this.transitionT0('index');
+      }
     }
+
+    // update(question, params) {
+    //   Object.keys(params).forEach(function(key) {
+    //     if(params[key]!==undefined) {
+    //       question.set(key,params[key]);
+    //     }
+    //   });
+    //   question.save();
+    //   this.transitionTo('index');
+    // },
+
+    // destroyQuestion(question) {
+    //   question.destroyRecord();
+    //   this.transitionTo('index');
+    // }
   }
 });
